# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para acompanhamento de pacientes tabagistas na APS/SUS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team (must have no patients or transfer them first)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Team has patients; transfer them before deleting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{id}/transfer-patients:
    post:
      operationId: transferTeamPatients
      tags: [teams]
      summary: Transfer all patients from one team to another
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferPatientsInput"
      responses:
        "200":
          description: Patients transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients with optional filters
      parameters:
        - name: teamId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: microarea
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: alertType
          in: query
          required: false
          schema:
            type: ["string", "null"]
            enum: ["red", "yellow", "none", null]
        - name: sex
          in: query
          required: false
          schema:
            type: ["string", "null"]
            enum: ["masculino", "feminino", null]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/import:
    post:
      operationId: importPatients
      tags: [patients]
      summary: Import patients from CSV text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPatientsInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPatientsResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Team:
      type: object
      required: [id, name, patientCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        patientCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    TransferPatientsInput:
      type: object
      required: [targetTeamId]
      properties:
        targetTeamId:
          type: integer
    Patient:
      type: object
      required:
        - id
        - identification
        - age
        - sex
        - teamId
        - teamName
        - microarea
        - smokingStatus
        - hasOralLesion
        - diagnosis
        - patientStatus
        - alertLevel
        - registrationDate
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        identification:
          type: string
        age:
          type: integer
        sex:
          type: string
          enum: [masculino, feminino]
        teamId:
          type: integer
        teamName:
          type: string
        microarea:
          type: string
        smokingStatus:
          type: string
          enum: [ativo, ex-tabagista]
        hasOralLesion:
          type: boolean
        lesionType:
          type: ["string", "null"]
        diagnosis:
          type: string
          enum: [nenhum, em_investigacao, confirmado]
        lastEvaluationDate:
          type: ["string", "null"]
          format: date
        registrationDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        patientStatus:
          type: string
          enum: [ativo, inativo]
        alertLevel:
          type: string
          enum: [red, yellow, none]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PatientInput:
      type: object
      required:
        - identification
        - age
        - sex
        - teamId
        - microarea
        - smokingStatus
        - hasOralLesion
        - diagnosis
        - registrationDate
        - patientStatus
      properties:
        identification:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 0
        sex:
          type: string
          enum: [masculino, feminino]
        teamId:
          type: integer
        microarea:
          type: string
          minLength: 1
        smokingStatus:
          type: string
          enum: [ativo, ex-tabagista]
        hasOralLesion:
          type: boolean
        lesionType:
          type: ["string", "null"]
        diagnosis:
          type: string
          enum: [nenhum, em_investigacao, confirmado]
        lastEvaluationDate:
          type: ["string", "null"]
          format: date
        registrationDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        patientStatus:
          type: string
          enum: [ativo, inativo]
    PatientUpdate:
      type: object
      properties:
        identification:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 0
        sex:
          type: string
          enum: [masculino, feminino]
        teamId:
          type: integer
        microarea:
          type: string
          minLength: 1
        smokingStatus:
          type: string
          enum: [ativo, ex-tabagista]
        hasOralLesion:
          type: boolean
        lesionType:
          type: ["string", "null"]
        diagnosis:
          type: string
          enum: [nenhum, em_investigacao, confirmado]
        lastEvaluationDate:
          type: ["string", "null"]
          format: date
        registrationDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        patientStatus:
          type: string
          enum: [ativo, inativo]
    ImportPatientsInput:
      type: object
      required: [csvText]
      properties:
        csvText:
          type: string
          description: Raw CSV content with header row
    ImportRowError:
      type: object
      required: [line, message]
      properties:
        line:
          type: integer
        message:
          type: string
    ImportPatientsResult:
      type: object
      required: [inserted, skipped, errors]
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportRowError"
...