# lib/api-zod/src/generated/types/importRowError.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API para acompanhamento de pacientes tabagistas na APS/SUS
 * OpenAPI spec version: 0.1.0
 */

export interface ImportRowError {
  line: number;
  message: string;
}