# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API para acompanhamento de pacientes tabagistas na APS/SUS
 * OpenAPI spec version: 0.1.0
 */
import type { PatientAlertLevel } from './patientAlertLevel';
import type { PatientDiagnosis } from './patientDiagnosis';
import type { PatientPatientStatus } from './patientPatientStatus';
import type { PatientSex } from './patientSex';
import type { PatientSmokingStatus } from './patientSmokingStatus';

export interface Patient {
  id: number;
  identification: string;
  age: number;
  sex: PatientSex;
  teamId: number;
  teamName: string;
  microarea: string;
  smokingStatus: PatientSmokingStatus;
  hasOralLesion: boolean;
  /** @nullable */
  lesionType?: string | null;
  diagnosis: PatientDiagnosis;
  /** @nullable */
  lastEvaluationDate?: Date | null;
  registrationDate: Date;
  /** @nullable */
  notes?: string | null;
  patientStatus: PatientPatientStatus;
  alertLevel: PatientAlertLevel;
  createdAt: Date;
  updatedAt: Date;
}